csrf3

+++ backend/src/models/product.ts
@@ -30,9 +30,12 @@ const cardsSchema = new mongoose.Schema<IProduct>(
             fileName: {
                 type: String,
                 required: [true, 'Поле "image.fileName" должно быть заполнено'],
+                set: sanitizeText,
             },
-            originalName: String,
-            set: sanitizeText,
+            originalName: {
+                type: String,
+                set: sanitizeText,
+            },
         },
         category: {
             type: String,
@@ -53,11 +56,11 @@ const cardsSchema = new mongoose.Schema<IProduct>(
 
 cardsSchema.index({ title: 'text' })
 
-// Можно лучше: удалять старое изображением перед обновлением сущности
 cardsSchema.pre('findOneAndUpdate', async function deleteOldImage() {
     // @ts-ignore
     const updateImage = this.getUpdate().$set?.image
     const docToUpdate = await this.model.findOne(this.getQuery())
+
     if (updateImage && docToUpdate) {
         unlink(
             join(__dirname, `../public/${docToUpdate.image.fileName}`),
@@ -66,11 +69,12 @@ cardsSchema.pre('findOneAndUpdate', async function deleteOldImage() {
     }
 })
 
-// Можно лучше: удалять файл с изображением после удаление сущности
 cardsSchema.post('findOneAndDelete', async (doc: IProduct) => {
-    unlink(join(__dirname, `../public/${doc.image.fileName}`), (err) =>
-        console.log(err)
-    )
+    if (doc?.image?.fileName) {
+        unlink(join(__dirname, `../public/${doc.image.fileName}`), (err) =>
+            console.log(err)
+        )
+    }
 })
 
 export default mongoose.model<IProduct>('product', cardsSchema)
